media_modules: t7 bringup code back to mainline. [1/1]
PD#SWPL-36783

Problem:
decoder t7 bringup.

Solution:
t7 media_modules code back to mainline.
ucode commit id: 75a2cf2
ucode change id: I2a17b0d
ucode gerrit id: 132514
note: t7 ucode has been included in last version,
so do not need update.

Verify:
an400

+++ drivers/stream_input/subtitle/subtitle.c
@@ -338,7 +338,7 @@ static ssize_t store_data(struct class *class, struct class_attribute *attr,
 	ssize_t r;
 
 	r = kstrtoint(buf, 0, &address);
-	if ((r == 0))
+	if (r == (ssize_t)0)
 		return -EINVAL;
 #if 0
 	if (subtitle_data[subtitle_write_pos].subtitle_size > 0) {
